Revert "refactor: move comment directive logic."
