+++ crates/tombi-lsp/src/completion/value/local_date.rs
@@ -1,13 +1,14 @@
-use tombi_comment_directive::{
-    get_value_comment_directive_content_with_schema_uri, value::LocalDateCommonRules,
-};
+use tombi_comment_directive::value::LocalDateCommonRules;
 use tombi_extension::CompletionKind;
 use tombi_future::Boxable;
 use tombi_schema_store::{Accessor, CurrentSchema, LocalDateSchema, SchemaUri};
 
-use crate::completion::{
-    comment::get_tombi_comment_directive_content_completion_contents, CompletionContent,
-    CompletionEdit, CompletionHint, FindCompletionContents,
+use crate::{
+    comment_directive::get_value_comment_directive_content_with_schema_uri,
+    completion::{
+        comment::get_tombi_comment_directive_content_completion_contents, CompletionContent,
+        CompletionEdit, CompletionHint, FindCompletionContents,
+    },
 };
 
 impl FindCompletionContents for tombi_document_tree::LocalDate {
