+++ crates/tombi-lsp/src/goto_type_definition/value/array.rs
@@ -2,14 +2,16 @@ use std::borrow::Cow;
 
 use itertools::Itertools;
 
-use tombi_comment_directive::get_array_comment_directive_content_with_schema_uri;
 use tombi_future::Boxable;
 use tombi_schema_store::{Accessor, ArraySchema, CurrentSchema, DocumentSchema, ValueSchema};
 
-use crate::goto_type_definition::{
-    all_of::get_all_of_type_definition, any_of::get_any_of_type_definition,
-    comment::get_tombi_value_comment_directive_type_definition, one_of::get_one_of_type_definition,
-    GetTypeDefinition, TypeDefinition,
+use crate::{
+    comment_directive::get_array_comment_directive_content_with_schema_uri,
+    goto_type_definition::{
+        all_of::get_all_of_type_definition, any_of::get_any_of_type_definition,
+        comment::get_tombi_value_comment_directive_type_definition,
+        one_of::get_one_of_type_definition, GetTypeDefinition, TypeDefinition,
+    },
 };
 
 impl GetTypeDefinition for tombi_document_tree::Array {
