+++ crates/tombi-lsp/src/hover/comment.rs
@@ -1,13 +1,16 @@
 use tombi_comment_directive::{
-    document::TombiDocumentDirectiveContent, CommentDirectiveContext, GetCommentDirectiveContext,
-    TombiCommentDirectiveImpl, TOMBI_COMMENT_DIRECTIVE_TOML_VERSION,
+    document::TombiDocumentDirectiveContent, TombiCommentDirectiveImpl,
+    TOMBI_COMMENT_DIRECTIVE_TOML_VERSION,
 };
 use tombi_comment_directive_store::comment_directive_document_schema;
 use tombi_document_tree::IntoDocumentTreeAndErrors;
 use tombi_uri::SchemaUri;
 
 use crate::{
-    comment_directive::{VALUE_TOMBI_DIRECTIVE_DESCRIPTION, VALUE_TOMBI_DIRECTIVE_TITLE},
+    comment_directive::{
+        CommentDirectiveContext, GetCommentDirectiveContext, VALUE_TOMBI_DIRECTIVE_DESCRIPTION,
+        VALUE_TOMBI_DIRECTIVE_TITLE,
+    },
     handler::get_hover_keys_with_range,
     hover::{get_hover_content, HoverContent, HoverDirectiveContent},
     DOCUMENT_SCHEMA_DIRECTIVE_DESCRIPTION, DOCUMENT_SCHEMA_DIRECTIVE_TITLE,
