+++ crates/tombi-lsp/src/hover/value/array.rs
@@ -2,13 +2,13 @@ use std::borrow::Cow;
 
 use itertools::Itertools;
 
-use tombi_comment_directive::get_array_comment_directive_content_with_schema_uri;
 use tombi_future::Boxable;
 use tombi_schema_store::{
     Accessor, Accessors, ArraySchema, CurrentSchema, DocumentSchema, ValueSchema, ValueType,
 };
 
 use crate::{
+    comment_directive::get_array_comment_directive_content_with_schema_uri,
     hover::{
         all_of::get_all_of_hover_content,
         any_of::get_any_of_hover_content,
