+++ crates/tombi-lsp/src/hover/value/offset_date_time.rs
@@ -1,9 +1,8 @@
-use tombi_comment_directive::{
-    get_value_comment_directive_content_with_schema_uri, value::OffsetDateTimeCommonRules,
-};
+use tombi_comment_directive::value::OffsetDateTimeCommonRules;
 use tombi_schema_store::{Accessor, CurrentSchema, OffsetDateTimeSchema, ValueSchema};
 
 use crate::{
+    comment_directive::get_value_comment_directive_content_with_schema_uri,
     hover::{
         all_of::get_all_of_hover_content,
         any_of::get_any_of_hover_content,
